# contracts/trader-console/openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "additionalProperties": false,
        "description": "Stable error envelope for resource failures.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiError",
        "type": "object"
      },
      "BacktestAssumptions": {
        "additionalProperties": false,
        "description": "Bounded execution and missing-data assumptions for preflight.",
        "properties": {
          "allow_latest_prior_bar": {
            "default": true,
            "title": "Allow Latest Prior Bar",
            "type": "boolean"
          },
          "allow_price_carry_forward": {
            "default": true,
            "title": "Allow Price Carry Forward",
            "type": "boolean"
          },
          "fee_bps": {
            "default": 0.0,
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Fee Bps",
            "type": "number"
          },
          "fee_fixed_per_order": {
            "default": 0.0,
            "maximum": 1000000.0,
            "minimum": 0.0,
            "title": "Fee Fixed Per Order",
            "type": "number"
          },
          "fee_minimum": {
            "default": 0.0,
            "maximum": 1000000.0,
            "minimum": 0.0,
            "title": "Fee Minimum",
            "type": "number"
          },
          "fill_model": {
            "default": "full_fill",
            "enum": [
              "full_fill",
              "next_bar"
            ],
            "title": "Fill Model",
            "type": "string"
          },
          "latency_ms": {
            "default": 0,
            "maximum": 60000.0,
            "minimum": 0.0,
            "title": "Latency Ms",
            "type": "integer"
          },
          "slippage_bps": {
            "default": 0.0,
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Slippage Bps",
            "type": "number"
          }
        },
        "title": "BacktestAssumptions",
        "type": "object"
      },
...
      "BacktestDefinitionRevision": {
        "additionalProperties": false,
        "description": "Immutable Console-owned definition revision returned after persistence.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "definition": {
            "$ref": "#/components/schemas/BacktestDefinition"
          },
          "definition_id": {
            "title": "Definition Id",
            "type": "string"
          },
          "definition_version": {
            "const": 1,
            "default": 1,
            "title": "Definition Version",
            "type": "integer"
          },
          "fingerprint": {
            "maxLength": 64,
            "minLength": 64,
            "pattern": "^[0-9a-f]{64}$",
            "title": "Fingerprint",
            "type": "string"
          },
          "revision": {
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          },
          "scope_id": {
            "title": "Scope Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "definition_id",
          "scope_id",
          "revision",
          "fingerprint",
          "definition",
          "created_at",
          "updated_at"
        ],
        "title": "BacktestDefinitionRevision",
        "type": "object"
      },
      "BacktestDefinitionsResponse": {
        "additionalProperties": false,
        "description": "Bounded page of the latest immutable revision for each definition.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BacktestDefinitionRevision"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          }
        },
        "required": [
          "items",
          "page"
        ],
        "title": "BacktestDefinitionsResponse",
        "type": "object"
      },
...
      "BacktestExecutionSubmit": {
        "additionalProperties": false,
        "description": "Idempotent request to enqueue one persisted definition revision.",
        "properties": {
          "definition_id": {
            "format": "uuid",
            "title": "Definition Id",
            "type": "string"
          },
          "idempotency_key": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Idempotency Key",
            "type": "string"
          }
        },
        "required": [
          "definition_id",
          "idempotency_key"
        ],
        "title": "BacktestExecutionSubmit",
        "type": "object"
      },
      "BacktestExecutionsResponse": {
        "additionalProperties": false,
        "description": "Bounded page of durable execution command records.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BacktestExecutionRecord"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          }
        },
        "required": [
          "items",
          "page"
        ],
        "title": "BacktestExecutionsResponse",
        "type": "object"
      },
      "BacktestPreflightRequest": {
        "additionalProperties": false,
        "description": "User-authored draft validated before persistence or execution.",
        "properties": {
          "asset_class": {
            "title": "Asset Class",
            "type": "string"
          },
          "assumptions": {
            "$ref": "#/components/schemas/BacktestAssumptions"
          },
          "benchmark_id": {
            "default": "buy_hold",
            "enum": [
              "buy_hold",
              "none"
            ],
            "title": "Benchmark Id",
            "type": "string"
          },
          "display_name": {
            "default": "Untitled backtest",
            "maxLength": 120,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "end": {
            "format": "date-time",
            "title": "End",
            "type": "string"
          },
          "initial_cash": {
            "default": 100000.0,
            "maximum": 1000000000000.0,
            "minimum": 0.0,
            "title": "Initial Cash",
            "type": "number"
          },
          "initial_positions": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/InitialPositionInput"
            },
            "title": "Initial Positions",
            "type": "array"
          },
          "resource_limits": {
            "$ref": "#/components/schemas/BacktestResourceLimits"
          },
          "risk_catalogue_version": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Catalogue Version"
          },
          "risk_parameters": {
            "additionalProperties": true,
            "title": "Risk Parameters",
            "type": "object"
          },
          "risk_profile_id": {
            "default": "noop",
            "maxLength": 100,
            "minLength": 1,
            "title": "Risk Profile Id",
            "type": "string"
          },
          "start": {
            "format": "date-time",
            "title": "Start",
            "type": "string"
          },
          "strategy_catalogue_version": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy Catalogue Version"
          },
          "strategy_parameters": {
            "additionalProperties": true,
            "title": "Strategy Parameters",
            "type": "object"
          },
          "strategy_profile_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Strategy Profile Id",
            "type": "string"
          },
          "symbols": {
            "items": {
              "type": "string"
            },
            "maxItems": 50,
            "minItems": 1,
            "title": "Symbols",
            "type": "array"
          },
          "timeframe": {
            "title": "Timeframe",
            "type": "string"
          }
        },
        "required": [
          "strategy_profile_id",
          "asset_class",
          "symbols",
          "timeframe",
          "start",
          "end"
        ],
        "title": "BacktestPreflightRequest",
        "type": "object"
      },
      "BacktestPreflightResponse": {
        "additionalProperties": false,
        "description": "Normalized preflight result with no producer or command side effects.",
        "properties": {
          "catalogue_version": {
            "title": "Catalogue Version",
            "type": "string"
          },
          "coverage": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/BacktestCoverageCheck"
            },
            "title": "Coverage",
            "type": "array"
          },
          "definition_fingerprint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Definition Fingerprint"
          },
          "issues": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/PreflightIssue"
            },
            "title": "Issues",
            "type": "array"
          },
          "normalized_definition": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BacktestDefinition"
              },
              {
                "type": "null"
              }
            ]
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          }
        },
        "required": [
          "valid",
          "catalogue_version"
        ],
        "title": "BacktestPreflightResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InitialPositionInput": {
        "additionalProperties": false,
        "description": "Optional initial position included in a typed backtest definition.",
        "properties": {
          "avg_price": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Price"
          },
          "qty": {
            "title": "Qty",
            "type": "number"
          },
          "symbol": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "qty"
        ],
        "title": "InitialPositionInput",
        "type": "object"
      },
...
      "PageInfo": {
        "additionalProperties": false,
        "description": "Bounded pagination evidence returned with collection resources.",
        "properties": {
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "limit": {
            "maximum": 50000.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "limit",
          "offset",
          "total",
          "has_more"
        ],
        "title": "PageInfo",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Trader Console API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/backtests/definitions": {
      "get": {
        "description": "List latest immutable revisions in the configured scope.",
        "operationId": "list_backtest_definitions_api_backtests_definitions_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestDefinitionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "List Backtest Definitions",
        "tags": [
          "backtests"
        ]
      },
      "post": {
        "description": "Preflight and persist the first immutable revision for a definition.",
        "operationId": "create_backtest_definition_api_backtests_definitions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BacktestPreflightRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestDefinitionRevision"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestPreflightResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Create Backtest Definition",
        "tags": [
          "backtests"
        ]
      }
    },
    "/api/backtests/definitions/{definition_id}": {
      "get": {
        "description": "Return the latest immutable revision for one definition identity.",
        "operationId": "get_backtest_definition_api_backtests_definitions__definition_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "definition_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Definition Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestDefinitionRevision"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Backtest Definition",
        "tags": [
          "backtests"
        ]
      }
    },
    "/api/backtests/definitions/{definition_id}/revisions": {
      "post": {
        "description": "Preflight and append one immutable revision to an existing definition.",
        "operationId": "create_backtest_definition_revision_api_backtests_definitions__definition_id__revisions_post",
        "parameters": [
          {
            "in": "path",
            "name": "definition_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Definition Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BacktestPreflightRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestDefinitionRevision"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestPreflightResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Create Backtest Definition Revision",
        "tags": [
          "backtests"
        ]
      }
    },
    "/api/backtests/executions": {
      "get": {
        "description": "List durable commands in the configured scope.",
        "operationId": "list_backtest_executions_api_backtests_executions_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestExecutionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "List Backtest Executions",
        "tags": [
          "backtests"
        ]
      },
      "post": {
        "description": "Submit one idempotent queued command for an immutable definition.",
        "operationId": "submit_backtest_execution_api_backtests_executions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BacktestExecutionSubmit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestExecutionRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Submit Backtest Execution",
        "tags": [
          "backtests"
        ]
      }
    },
    "/api/backtests/executions/{execution_id}": {
      "get": {
        "description": "Return durable status for one execution command.",
        "operationId": "get_backtest_execution_api_backtests_executions__execution_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "execution_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Execution Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestExecutionRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Backtest Execution",
        "tags": [
          "backtests"
        ]
      }
    },
...
  }
}